Improve structure and documentation of Filter Comparison connector
